fixed and tested comment delete

+++ react-app/src/components/CommentCard/index.js
@@ -11,9 +11,12 @@ import './CommentCard.css'
 export default function CommentCard({ post, comment }) {
 	const dispatch = useDispatch();
 	let sessionUser = useSelector(state => state.session.user);
+	
 
 	const deleteComment = (commentId) => {
-		dispatch(postsActions.deleteComment(commentId, post.id))
+		if (window.confirm('Are you sure you would like to delete your comment?')) {
+			dispatch(postsActions.deleteComment(commentId, post.id))
+		}
 	}
 
 	return (
@@ -32,7 +35,7 @@ export default function CommentCard({ post, comment }) {
 
 			{comment.user.id === sessionUser.id &&
 				<div className='edit-and-delete-buttons'>
-					<FontAwesomeIcon icon={faEdit} id='edit-comment-button' onClick={() => deleteComment(comment.id)} />
+					<FontAwesomeIcon icon={faEdit} id='edit-comment-button' onClick={() => alert('Sorry! This feature is under construction')} />
 					<FontAwesomeIcon icon={faTrash} id='delete-comment-button' onClick={() => deleteComment(comment.id)} />
 				</div>
 			}
